refactor(Microsoft Excel SharePoint Node): Move single-use types next to their code (no-changelog)
Fold interfaces.ts down to just AuthContext, the one type used across
nearly every helper. Move each other type next to its one real
consumer instead: ExcelSharePointCredentialType into constants.ts,
GraphRequestError into converters.ts, and ErrorMapper into
errorMappers.ts. Also fold getExcelSharePointCredentialType back into
transport/index.ts, since nothing outside that file uses it yet.

Addresses Stephen's review feedback on PR #34097: prefer a type next
to the code it describes over a shared types file, unless that type
is genuinely reused across the codebase.
