+++ packages/nodes-base/nodes/Microsoft/ExcelSharePoint/helpers/interfaces.ts
@@ -1,19 +1,6 @@
-import type { IExecuteFunctions, ILoadOptionsFunctions, NodeApiError } from 'n8n-workflow';
+import type { IExecuteFunctions, ILoadOptionsFunctions } from 'n8n-workflow';
 
-import type { SERVICE_PRINCIPAL_AUTH } from './constants';
-
-export type ExcelSharePointCredentialType = 'microsoftOAuth2Api' | typeof SERVICE_PRINCIPAL_AUTH;
-
-export type GraphRequestError = {
-	httpCode?: string | number | null;
-	statusCode?: string | number | null;
-	message?: string;
-	code?: string;
-	error?: { error?: GraphRequestError };
-};
-
+// Used as `this` by every helper in this node (credential resolution, error
+// mapping, the request itself), so it earns a shared home unlike the other
+// types here, which each live next to their one real consumer.
 export type AuthContext = IExecuteFunctions | ILoadOptionsFunctions;
-
-// Both sign-in methods now go through the same httpRequestWithAuthentication
-// call; only the shape of the resulting error still differs by credential type.
-export type ErrorMapper = (this: AuthContext, error: GraphRequestError) => NodeApiError;
