+++ packages/nodes-base/nodes/Microsoft/ExcelSharePoint/transport/index.ts
@@ -1,9 +1,17 @@
 import type { IDataObject, IHttpRequestMethods, IHttpRequestOptions } from 'n8n-workflow';
 
+import { SERVICE_PRINCIPAL_AUTH, type ExcelSharePointCredentialType } from '../helpers/constants';
 import { buildRequestOptions } from '../helpers/converters';
-import { getExcelSharePointCredentialType } from '../helpers/credentials';
 import { getErrorMapper } from '../helpers/errorMappers';
 import type { AuthContext } from '../helpers/interfaces';
+
+// Not expected to be reused outside this file yet; move it back to its own
+// helpers/credentials.ts if a second consumer (e.g. a load-options dropdown) needs it.
+export function getExcelSharePointCredentialType(this: AuthContext): ExcelSharePointCredentialType {
+	// In load-options contexts the 2nd arg is the fallback, not an item index — keep the 2-arg form
+	const selected = this.getNodeParameter('authentication', 0);
+	return selected === SERVICE_PRINCIPAL_AUTH ? SERVICE_PRINCIPAL_AUTH : 'microsoftOAuth2Api';
+}
 
 export async function microsoftApiRequest<T extends IDataObject = IDataObject>(
 	this: AuthContext,
